QUALITY: Move api url to Angular
