+++ angular.json
@@ -288,7 +288,6 @@
               "build/src/app/api/models/teaching-period.js",
               "build/src/app/api/models/group.js",
               "build/src/app/api/models/unit.js",
-              "build/src/app/api/api-url.js",
               "build/src/app/api/api.js",
               "build/src/app/api/resource-plus.js",
               "build/src/app/errors/errors.js",
