+++ src/app/api/api.coffee
@@ -1,5 +1,4 @@
 angular.module("doubtfire.api", [
-  "doubtfire.api.api-url"
   "doubtfire.api.resource-plus"
   "doubtfire.api.models"
 ])
