+++ src/app/api/models/break.coffee
@@ -1,6 +1,6 @@
 angular.module("doubtfire.api.models.break", [])
 
-.factory("Break", (resourcePlus, currentUser, api) ->
+.factory("Break", (resourcePlus) ->
   Break = resourcePlus "/teaching_periods/:teaching_period_id/breaks", { teaching_period_id: "@teaching_period_id" }
   Break
 )
