+++ src/app/api/models/intended-learning-outcome.coffee
@@ -1,10 +1,10 @@
 angular.module("doubtfire.api.models.intended-learning-outcome", [])
 
-.factory("IntendedLearningOutcome", (resourcePlus, api, currentUser) ->
+.factory("IntendedLearningOutcome", (resourcePlus, DoubtfireConstants, currentUser) ->
   IntendedLearningOutcome = resourcePlus "/units/:unit_id/outcomes/:id", { id: "@id", unit_id: "@unit_id" }
 
   IntendedLearningOutcome.getOutcomeBatchUploadUrl = (unit) ->
-    "#{api}/units/#{unit.id}/outcomes/csv?auth_token=#{currentUser.authenticationToken}"
+    "#{DoubtfireConstants.API_URL}/units/#{unit.id}/outcomes/csv?auth_token=#{currentUser.authenticationToken}"
 
   IntendedLearningOutcome
 )
