+++ src/app/api/models/rollover-unit.coffee
@@ -1,6 +1,6 @@
 angular.module("doubtfire.api.models.rollover-unit", [])
 
-.factory("RolloverUnit", (resourcePlus, api, currentUser) ->
+.factory("RolloverUnit", (resourcePlus) ->
   DuplicatedUnit = resourcePlus "/units/:id/rollover", { id: "@id" }
 
   return DuplicatedUnit
