+++ src/app/api/models/task-completion-csv.coffee
@@ -1,15 +1,15 @@
 angular.module("doubtfire.api.models.task-completion-csv", [])
 
-.service("TaskCompletionCsv", (api, $window, currentUser) ->
+.service("TaskCompletionCsv", (DoubtfireConstants, $window, currentUser) ->
   this.downloadFile = (unit) ->
-    $window.open "#{api}/csv/units/#{unit.id}/task_completion.json?auth_token=#{currentUser.authenticationToken}", "_blank"
+    $window.open "#{DoubtfireConstants.API_URL}/csv/units/#{unit.id}/task_completion.json?auth_token=#{currentUser.authenticationToken}", "_blank"
 
   return this
 )
 
-.service("TutorAssessmentCsv", (api, $window, currentUser) ->
+.service("TutorAssessmentCsv", (DoubtfireConstants, $window, currentUser) ->
   this.downloadFile = (unit) ->
-    $window.open "#{api}/csv/units/#{unit.id}/tutor_assessments.json?auth_token=#{currentUser.authenticationToken}", "_blank"
+    $window.open "#{DoubtfireConstants.API_URL}/csv/units/#{unit.id}/tutor_assessments.json?auth_token=#{currentUser.authenticationToken}", "_blank"
 
   return this
 )
