+++ src/app/config/privacy-policy/privacy-policy.coffee
@@ -1,13 +1,13 @@
 angular.module("doubtfire.config.privacy-policy", [])
 
-.factory('PrivacyPolicy', ($http, api) ->
+.factory('PrivacyPolicy', ($http, DoubtfireConstants) ->
   privacyPolicy = {
     privacy: '',
     plagiarism: '',
     loaded: false,
   }
 
-  $http.get("#{api}/settings/privacy").then ((response) ->
+  $http.get("#{DoubtfireConstants.API_URL}/settings/privacy").then ((response) ->
     privacyPolicy.privacy = response.data.privacy
     privacyPolicy.plagiarism = response.data.plagiarism
     privacyPolicy.loaded = true
