+++ src/app/sessions/auth/http-auth-injector.coffee
@@ -4,13 +4,13 @@ angular.module("doubtfire.sessions.auth.http-auth-injector", [])
 # all
 #
 .config(($httpProvider) ->
-  $httpProvider.interceptors.push ($q, $rootScope, api, currentUser) ->
+  $httpProvider.interceptors.push ($q, $rootScope, DoubtfireConstants, currentUser) ->
     #
     # Inject authentication token for requests
     #
     injectAuthForRequest = (request) ->
       # Intercept API requests and inject the auth token.
-      if _.startsWith(request.url, api) and currentUser.authenticationToken?
+      if _.startsWith(request.url, DoubtfireConstants.API_URL) and currentUser.authenticationToken?
         request.params = {} unless _.has request, "params"
         request.params.auth_token = currentUser.authenticationToken
       request or $q.when request
@@ -20,7 +20,7 @@ angular.module("doubtfire.sessions.auth.http-auth-injector", [])
     #
     injectAuthForResponseWithError = (response) ->
       # Intercept unauthorised API responses and fire an event.
-      if response.config && response.config.url and _.startsWith(response.config.url, api)
+      if response.config && response.config.url and _.startsWith(response.config.url, DoubtfireConstants.API_URL)
         # Timeout?
         if response.status is 419
           $rootScope.$broadcast "tokenTimeout"
