+++ src/app/sessions/states/sign-out/sign-out.coffee
@@ -14,14 +14,12 @@ angular.module("doubtfire.sessions.states.sign-out", [])
       pageTitle: "_Sign Out_"
   $stateProvider.state "sign_out", signOutStateData
 )
-.controller("SignOutCtrl", ($state, $timeout, $http, auth, api, currentUser) ->
-  if auth.signOut api + "/auth/" + currentUser.authenticationToken + ".json"
-    $http.get("#{api}/auth/signout_url").then ((response) ->
-      signoutURL = response.data.auth_signout_url || false
-      if signoutURL
-        window.location.assign(signoutURL)
-      else
-        $timeout (-> $state.go "sign_in"), 750
-    )
-  return this
+.controller("SignOutCtrl", ($state, $timeout, $http, auth, DoubtfireConstants, currentUser) ->
+  if currentUser?.authenticationToken?
+    auth.signOut("#{DoubtfireConstants.API_URL}/auth/#{currentUser.authenticationToken}.json")
+
+  if DoubtfireConstants.SignoutURL?
+    window.location.assign(DoubtfireConstants.SignoutURL)
+  else
+    $timeout (-> $state.go "sign_in"), 500
 )
